pyproject.toml
[build-system]
requires      = ["setuptools>=69.1.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "actirepo"
version = "0.0.1"
description = "Moodle activities repo organizer"
readme = "README.md"
authors = [
    { name = "fvarrui", email = "francisco.vargas.ruiz@gmail.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["activity", "repo", "moodle", "questions", "organizer"]
dependencies = [
    "Jinja2>=3.1.2",
    "Pillow>=10.1.0",
    "html2image>=2.0.5",
    "beautifulsoup4>=4.12.3",
    "tabulate>=0.9.0",
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/teuton-software/actirepo"

[project.scripts]
actirepo = "actirepo.__main__:main"
test = "cli.__main__:main"